+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: akant <akant@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/08 17:04:34 by akant         #+#    #+#                 */
-/*   Updated: 2020/11/08 17:10:23 by akant         ########   odam.nl         */
+/*   Updated: 2020/11/19 15:00:15 by alexanderka   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *string, int a, size_t len)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	while (i < len)
